Promote FFT solver pack/unpack helpers into a shared header
Move the previously inlined pack / unpack / solver_send / solver_recv /
ViewAccess helpers from PoissonSolvers/FFTOpenPoissonSolver.h(pp) into a
new shared header Field/FieldBufferOps.hpp (namespace ippl::detail), so
other utilities can reuse the same device-parallel, CUDA-aware-MPI
packing primitives without duplicating them.

- FFTOpenPoissonSolver.h: include the new header, drop the local
  ViewAccess definition.
- FFTOpenPoissonSolver.hpp: drop the inlined helpers, pull them into
  namespace ippl via using-declarations so existing call sites remain
  unqualified.
- Add a value-preserving pack variant (pack_field / solver_send_field)
  that works for arbitrary element types (e.g. ippl::Vector) by copying
  view(i,j,k) directly instead of going through the Kokkos::complex /
  Kokkos::real round-trip of the legacy pack.
- Communicate/Tags.h: add a dedicated FIELD_MIRROR MPI tag (40500) for
  clients of the generic field-mirror utility built on top of these
  helpers.
