+++ src/PoissonSolvers/FFTOpenPoissonSolver.h
@@ -20,53 +20,13 @@
 
 #include "Communicate/Archive.h"
 #include "FFT/FFT.h"
+#include "Field/FieldBufferOps.hpp"
 #include "Field/HaloCells.h"
 #include "FieldLayout/FieldLayout.h"
 #include "Meshes/UniformCartesian.h"
 #include "Poisson.h"
 
 namespace ippl {
-    namespace detail {
-
-        /*!
-         * Access a view that either contains a vector field or a scalar field
-         * in such a way that the correct element access is determined at compile
-         * time, reducing the number of functions needed to achieve the same
-         * behavior for both kinds of fields
-         * @tparam tensorRank indicates whether scalar, vector, or matrix field
-         * @tparam - the view type
-         */
-        template <int tensorRank, typename>
-        struct ViewAccess;
-
-        template <typename View>
-        struct ViewAccess<2, View> {
-            KOKKOS_INLINE_FUNCTION constexpr static auto& get(View&& view, unsigned dim1,
-                                                              unsigned dim2, size_t i, size_t j,
-                                                              size_t k) {
-                return view(i, j, k)[dim1][dim2];
-            }
-        };
-
-        template <typename View>
-        struct ViewAccess<1, View> {
-            KOKKOS_INLINE_FUNCTION constexpr static auto& get(View&& view, unsigned dim1,
-                                                              [[maybe_unused]] unsigned dim2,
-                                                              size_t i, size_t j, size_t k) {
-                return view(i, j, k)[dim1];
-            }
-        };
-
-        template <typename View>
-        struct ViewAccess<0, View> {
-            KOKKOS_INLINE_FUNCTION constexpr static auto& get(View&& view,
-                                                              [[maybe_unused]] unsigned dim1,
-                                                              [[maybe_unused]] unsigned dim2,
-                                                              size_t i, size_t j, size_t k) {
-                return view(i, j, k);
-            }
-        };
-    }  // namespace detail
 
     template <typename FieldLHS, typename FieldRHS>
     class FFTOpenPoissonSolver : public Poisson<FieldLHS, FieldRHS> {
